finish splitting interfaces into their respective packages
