+++ sia/host.go
@@ -1,8 +1,5 @@
 package sia
 
-type Host interface {
-}
-
 /*
 import (
 	"errors"
